style: rename fuzzymatch
