+++ boaviztapi/utils/fuzzymatch.py
@@ -1,6 +1,6 @@
-from typing import Optional
-
+import pandas as pd
 from pandas import DataFrame
+from pandas.core.series import Series
 from rapidfuzz import process, fuzz
 
 
@@ -11,3 +11,11 @@ def fuzzymatch_attr_from_pdf(name: str, attr: str, pdf: DataFrame) -> str:
     if result is not None:
         result = result[0] if result[1] > 88.0 else None
     return result or None
+
+
+def fuzzymatch(s: pd.Series, value: str, threshold: float = 90.) -> pd.Series:
+    return s.apply(lambda x: fuzz.token_sort_ratio(x.lower(), value.lower()) >= threshold)
+
+
+def pandas() -> None:
+    Series.fuzzymatch = fuzzymatch
